requirements.txt
Django==4.2.4
djangorestframework==3.14.0
psycopg2==2.9.7
django-cors-headers==3.14.0
django-redis==5.2.0
djangorestframework-simplejwt==5.2.2
redis==5.0.10
